7) Выделил общий    базовый  класс 8) Добавил  тест  по удалению  группы 9) Рефакторинг  кода.
