+++ README.md
@@ -11,4 +11,6 @@
 4) Установка address book
 5) Первый  тест для  книги
 # lesson 3 - p.6
-Добавлен  первый  АвтоТест
+1)Добавлен первый АвтоТест
+2)Провел рефакторинг создания групп
+3)Добавил обьект GroupDate
