+++ address-book-tests/src/test/java/ru/rt/client/GroupCreationTest.java
@@ -1,0 +1,16 @@
+package ru.rt.client;
+
+import org.testng.annotations.Test;
+
+public class GroupCreationTest extends TestBase {
+
+
+    @Test
+    public void testGroupCreation() throws Exception {
+        navToGroupPage();
+        initGroupCreation();
+        fillGroupForm(new GroupDate("group1", "header test", "group footer"));
+        submitGroupCreation();
+    }
+
+}
