+++ address-book-tests/src/test/java/ru/rt/client/GroupDeletionTest.java
@@ -1,0 +1,16 @@
+package ru.rt.client;
+
+import org.testng.annotations.Test;
+
+public class GroupDeletionTest extends TestBase {
+
+
+    @Test
+    public void testGroupDeletion() throws Exception {
+        navToGroupPage();
+        SelectGroup();
+        clickDeleteBtn();
+    }
+
+
+}
